fixes // so far everything compiles

+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: csilva-r <csilva-r@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/18 16:23:21 by csilva-r          #+#    #+#             */
-/*   Updated: 2024/04/18 16:32:10 by csilva-r         ###   ########.fr       */
+/*   Updated: 2024/04/19 14:12:52 by csilva-r         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s1[i] && s2[i] && i <= n)
